Dataクラスも分けたよ
