+++ app/src/main/java/com/example/tsundokun/ui/home/component/WebPage.kt
@@ -1,0 +1,19 @@
+package com.example.tsundokun.ui.home.component
+
+/*
+ * Webページの情報のデータクラス
+ */
+data class WebPage(
+    val title: String?,
+    val ogpImageUrl: String?,
+    val faviconImageUrl: String?,
+    val link: String?,
+)
+/*
+ * タブの種類
+ */
+enum class Screen {
+    ALL, FAVORITE
+}
+
+
